refactor: move loading in a different file
