+++ packages/cli/src/commands/mock.ts
@@ -1,10 +1,6 @@
-import { transformOas2Operation, transformOas3Operation } from '@stoplight/http-spec';
-import { parse } from '@stoplight/yaml';
-import * as fs from 'fs';
-import { flatten, get, keys, map } from 'lodash';
 import { CommandModule } from 'yargs';
-import { httpAndFileResolver } from '../resolvers/http-and-file';
 import { createMultiProcessPrism, CreatePrismOptions, createSingleProcessPrism } from '../util/createServer';
+import getHttpOperations from '../util/getHttpOperations';
 
 const mockCommand: CommandModule = {
   describe: 'Start a mock server with the given spec file',
@@ -16,24 +12,7 @@ const mockCommand: CommandModule = {
         type: 'string',
       })
       .middleware(async argv => {
-        const fileContent = fs.readFileSync(argv.spec!, { encoding: 'utf8' });
-        const parsedContent = parse(fileContent);
-        const { result: resolvedContent } = await httpAndFileResolver.resolve(parsedContent);
-
-        const transformFn = get(parsedContent, 'swagger') ? transformOas2Operation : transformOas3Operation;
-
-        const paths = keys(get(resolvedContent, 'paths'));
-        argv.operations = flatten(
-          map(paths, path =>
-            keys(get(resolvedContent, ['paths', path])).map(method =>
-              transformFn({
-                document: resolvedContent,
-                path,
-                method,
-              }),
-            ),
-          ),
-        );
+        argv.operations = await getHttpOperations(argv.spec!);
       })
       .fail(() => {
         yargs.showHelp();
